feat : CSR-only를 보장하기 위해 컴포넌트 새로 생성
