+++ app/auth/success/page.tsx
@@ -1,22 +1,12 @@
 "use client";
 
+import { Suspense } from "react";
+import SuccessHandler from "./success-handler";
 
-import {useEffect} from "react";
-import {useSearchParams,useRouter} from "next/navigation";
-
-
-export default function AuthSuccessPage(){
-    const params = useSearchParams();
-    const router = useRouter();
-
-    useEffect(() => {
-        const token = params.get("token");
-
-        if (token){
-            localStorage.setItem("accessToken",token);
-            router.replace("/"); //로그인 완료후 메인으로 이동
-        }
-    }, [params,router]);
-
-    return <p>로그인 처리중...</p>
+export default function AuthSuccessPage() {
+    return (
+        <Suspense fallback={<p>로그인 처리중...</p>}>
+            <SuccessHandler />
+        </Suspense>
+    );
 }
